% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bey_coachella_palette}
\alias{bey_coachella_palette}
\title{Use \code{\link{bey_coachella_palette}} to construct palettes of desired length.}
\usage{
bey_coachella_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colours desired. Unfortunately most palettes now only
have 4 or 5 colours. But hopefully we'll add more palettes soon. All colour
schemes are derived from the most excellent Tumblr blog:
\href{http://beyoncepalettes.tumblr.com/}{Beyoncé Palettes}.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}

\item{number}{ID of desired palette (1 to 3)}
}
\value{
A vector of colours.
}
\description{
Beyoncé Coachella 2019 colour palettes.
}
\examples{
bey_coachella_palette(2, 8)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- beyonce_palette(21, number = 2, type = "continuous")
image(volcano, col = pal)

}
\keyword{colours}
