% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colours.R
\name{beyonce_palette}
\alias{beyonce_palette}
\title{Beyoncé palette generator}
\usage{
beyonce_palette(number, n, type = c("discrete", "continuous"))
}
\arguments{
\item{number}{ID of desired palette (1 to 130)}

\item{n}{Number of colors desired. Unfortunately most palettes now only
have 4 or 5 colors. But hopefully we'll add more palettes soon. All color
schemes are derived from the most excellent Tumblr blog:
\href{http://beyoncepalettes.tumblr.com/}{Beyoncé Palettes}.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
A lot of Beyoncé color palettes.
}
\examples{
beyonce_palette(1, 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- beyonce_palette(21, number = 2, type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}

